+++ sdk/cdn/arm-cdn/samples/v9/javascript/afdCustomDomainsDeleteSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Deletes an existing AzureFrontDoor domain with the specified domain name under the specified subscription, resource group and profile.
  *
  * @summary Deletes an existing AzureFrontDoor domain with the specified domain name under the specified subscription, resource group and profile.
- * x-ms-original-file: specification/cdn/resource-manager/Microsoft.Cdn/stable/2023-05-01/examples/AFDCustomDomains_Delete.json
+ * x-ms-original-file: specification/cdn/resource-manager/Microsoft.Cdn/stable/2024-02-01/examples/AFDCustomDomains_Delete.json
  */
 async function afdCustomDomainsDelete() {
   const subscriptionId = process.env["CDN_SUBSCRIPTION_ID"] || "subid";