+++ sdk/cdn/arm-cdn/src/operations/afdOriginGroups.ts
@@ -16,7 +16,7 @@ import { CdnManagementClient } from "../cdnManagementClient";
 import {
   SimplePollerLike,
   OperationState,
-  createHttpPoller
+  createHttpPoller,
 } from "@azure/core-lro";
 import { createLroSpec } from "../lroImpl";
 import {
@@ -37,7 +37,7 @@ import {
   AfdOriginGroupsUpdateResponse,
   AfdOriginGroupsDeleteOptionalParams,
   AfdOriginGroupsListByProfileNextResponse,
-  AfdOriginGroupsListResourceUsageNextResponse
+  AfdOriginGroupsListResourceUsageNextResponse,
 } from "../models";
 
 /// <reference lib="esnext.asynciterable" />
@@ -63,12 +63,12 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
   public listByProfile(
     resourceGroupName: string,
     profileName: string,
-    options?: AfdOriginGroupsListByProfileOptionalParams
+    options?: AfdOriginGroupsListByProfileOptionalParams,
   ): PagedAsyncIterableIterator<AFDOriginGroup> {
     const iter = this.listByProfilePagingAll(
       resourceGroupName,
       profileName,
-      options
+      options,
     );
     return {
       next() {
@@ -85,9 +85,9 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
           resourceGroupName,
           profileName,
           options,
-          settings
+          settings,
         );
-      }
+      },
     };
   }
 
@@ -95,7 +95,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     options?: AfdOriginGroupsListByProfileOptionalParams,
-    settings?: PageSettings
+    settings?: PageSettings,
   ): AsyncIterableIterator<AFDOriginGroup[]> {
     let result: AfdOriginGroupsListByProfileResponse;
     let continuationToken = settings?.continuationToken;
@@ -103,7 +103,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
       result = await this._listByProfile(
         resourceGroupName,
         profileName,
-        options
+        options,
       );
       let page = result.value || [];
       continuationToken = result.nextLink;
@@ -115,7 +115,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         resourceGroupName,
         profileName,
         continuationToken,
-        options
+        options,
       );
       continuationToken = result.nextLink;
       let page = result.value || [];
@@ -127,12 +127,12 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
   private async *listByProfilePagingAll(
     resourceGroupName: string,
     profileName: string,
-    options?: AfdOriginGroupsListByProfileOptionalParams
+    options?: AfdOriginGroupsListByProfileOptionalParams,
   ): AsyncIterableIterator<AFDOriginGroup> {
     for await (const page of this.listByProfilePagingPage(
       resourceGroupName,
       profileName,
-      options
+      options,
     )) {
       yield* page;
     }
@@ -150,13 +150,13 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     originGroupName: string,
-    options?: AfdOriginGroupsListResourceUsageOptionalParams
+    options?: AfdOriginGroupsListResourceUsageOptionalParams,
   ): PagedAsyncIterableIterator<Usage> {
     const iter = this.listResourceUsagePagingAll(
       resourceGroupName,
       profileName,
       originGroupName,
-      options
+      options,
     );
     return {
       next() {
@@ -174,9 +174,9 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
           profileName,
           originGroupName,
           options,
-          settings
+          settings,
         );
-      }
+      },
     };
   }
 
@@ -185,7 +185,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     profileName: string,
     originGroupName: string,
     options?: AfdOriginGroupsListResourceUsageOptionalParams,
-    settings?: PageSettings
+    settings?: PageSettings,
   ): AsyncIterableIterator<Usage[]> {
     let result: AfdOriginGroupsListResourceUsageResponse;
     let continuationToken = settings?.continuationToken;
@@ -194,7 +194,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         resourceGroupName,
         profileName,
         originGroupName,
-        options
+        options,
       );
       let page = result.value || [];
       continuationToken = result.nextLink;
@@ -207,7 +207,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         profileName,
         originGroupName,
         continuationToken,
-        options
+        options,
       );
       continuationToken = result.nextLink;
       let page = result.value || [];
@@ -220,13 +220,13 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     originGroupName: string,
-    options?: AfdOriginGroupsListResourceUsageOptionalParams
+    options?: AfdOriginGroupsListResourceUsageOptionalParams,
   ): AsyncIterableIterator<Usage> {
     for await (const page of this.listResourceUsagePagingPage(
       resourceGroupName,
       profileName,
       originGroupName,
-      options
+      options,
     )) {
       yield* page;
     }
@@ -242,11 +242,11 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
   private _listByProfile(
     resourceGroupName: string,
     profileName: string,
-    options?: AfdOriginGroupsListByProfileOptionalParams
+    options?: AfdOriginGroupsListByProfileOptionalParams,
   ): Promise<AfdOriginGroupsListByProfileResponse> {
     return this.client.sendOperationRequest(
       { resourceGroupName, profileName, options },
-      listByProfileOperationSpec
+      listByProfileOperationSpec,
     );
   }
 
@@ -262,11 +262,11 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     originGroupName: string,
-    options?: AfdOriginGroupsGetOptionalParams
+    options?: AfdOriginGroupsGetOptionalParams,
   ): Promise<AfdOriginGroupsGetResponse> {
     return this.client.sendOperationRequest(
       { resourceGroupName, profileName, originGroupName, options },
-      getOperationSpec
+      getOperationSpec,
     );
   }
 
@@ -284,7 +284,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     profileName: string,
     originGroupName: string,
     originGroup: AFDOriginGroup,
-    options?: AfdOriginGroupsCreateOptionalParams
+    options?: AfdOriginGroupsCreateOptionalParams,
   ): Promise<
     SimplePollerLike<
       OperationState<AfdOriginGroupsCreateResponse>,
@@ -293,21 +293,20 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
   > {
     const directSendOperation = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ): Promise<AfdOriginGroupsCreateResponse> => {
       return this.client.sendOperationRequest(args, spec);
     };
     const sendOperationFn = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
+      let currentRawResponse: coreClient.FullOperationResponse | undefined =
+        undefined;
       const providedCallback = args.options?.onResponse;
       const callback: coreClient.RawResponseCallback = (
         rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
+        flatResponse: unknown,
       ) => {
         currentRawResponse = rawResponse;
         providedCallback?.(rawResponse, flatResponse);
@@ -316,8 +315,8 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         ...args,
         options: {
           ...args.options,
-          onResponse: callback
-        }
+          onResponse: callback,
+        },
       };
       const flatResponse = await directSendOperation(updatedArgs, spec);
       return {
@@ -325,8 +324,8 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         rawResponse: {
           statusCode: currentRawResponse!.status,
           body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
+          headers: currentRawResponse!.headers.toJSON(),
+        },
       };
     };
 
@@ -337,9 +336,9 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         profileName,
         originGroupName,
         originGroup,
-        options
+        options,
       },
-      spec: createOperationSpec
+      spec: createOperationSpec,
     });
     const poller = await createHttpPoller<
       AfdOriginGroupsCreateResponse,
@@ -347,7 +346,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     >(lro, {
       restoreFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs,
-      resourceLocationConfig: "azure-async-operation"
+      resourceLocationConfig: "azure-async-operation",
     });
     await poller.poll();
     return poller;
@@ -367,14 +366,14 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     profileName: string,
     originGroupName: string,
     originGroup: AFDOriginGroup,
-    options?: AfdOriginGroupsCreateOptionalParams
+    options?: AfdOriginGroupsCreateOptionalParams,
   ): Promise<AfdOriginGroupsCreateResponse> {
     const poller = await this.beginCreate(
       resourceGroupName,
       profileName,
       originGroupName,
       originGroup,
-      options
+      options,
     );
     return poller.pollUntilDone();
   }
@@ -393,7 +392,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     profileName: string,
     originGroupName: string,
     originGroupUpdateProperties: AFDOriginGroupUpdateParameters,
-    options?: AfdOriginGroupsUpdateOptionalParams
+    options?: AfdOriginGroupsUpdateOptionalParams,
   ): Promise<
     SimplePollerLike<
       OperationState<AfdOriginGroupsUpdateResponse>,
@@ -402,21 +401,20 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
   > {
     const directSendOperation = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ): Promise<AfdOriginGroupsUpdateResponse> => {
       return this.client.sendOperationRequest(args, spec);
     };
     const sendOperationFn = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
+      let currentRawResponse: coreClient.FullOperationResponse | undefined =
+        undefined;
       const providedCallback = args.options?.onResponse;
       const callback: coreClient.RawResponseCallback = (
         rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
+        flatResponse: unknown,
       ) => {
         currentRawResponse = rawResponse;
         providedCallback?.(rawResponse, flatResponse);
@@ -425,8 +423,8 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         ...args,
         options: {
           ...args.options,
-          onResponse: callback
-        }
+          onResponse: callback,
+        },
       };
       const flatResponse = await directSendOperation(updatedArgs, spec);
       return {
@@ -434,8 +432,8 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         rawResponse: {
           statusCode: currentRawResponse!.status,
           body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
+          headers: currentRawResponse!.headers.toJSON(),
+        },
       };
     };
 
@@ -446,9 +444,9 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         profileName,
         originGroupName,
         originGroupUpdateProperties,
-        options
+        options,
       },
-      spec: updateOperationSpec
+      spec: updateOperationSpec,
     });
     const poller = await createHttpPoller<
       AfdOriginGroupsUpdateResponse,
@@ -456,7 +454,7 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     >(lro, {
       restoreFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs,
-      resourceLocationConfig: "azure-async-operation"
+      resourceLocationConfig: "azure-async-operation",
     });
     await poller.poll();
     return poller;
@@ -476,14 +474,14 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     profileName: string,
     originGroupName: string,
     originGroupUpdateProperties: AFDOriginGroupUpdateParameters,
-    options?: AfdOriginGroupsUpdateOptionalParams
+    options?: AfdOriginGroupsUpdateOptionalParams,
   ): Promise<AfdOriginGroupsUpdateResponse> {
     const poller = await this.beginUpdate(
       resourceGroupName,
       profileName,
       originGroupName,
       originGroupUpdateProperties,
-      options
+      options,
     );
     return poller.pollUntilDone();
   }
@@ -500,25 +498,24 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     originGroupName: string,
-    options?: AfdOriginGroupsDeleteOptionalParams
+    options?: AfdOriginGroupsDeleteOptionalParams,
   ): Promise<SimplePollerLike<OperationState<void>, void>> {
     const directSendOperation = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ): Promise<void> => {
       return this.client.sendOperationRequest(args, spec);
     };
     const sendOperationFn = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
+      let currentRawResponse: coreClient.FullOperationResponse | undefined =
+        undefined;
       const providedCallback = args.options?.onResponse;
       const callback: coreClient.RawResponseCallback = (
         rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
+        flatResponse: unknown,
       ) => {
         currentRawResponse = rawResponse;
         providedCallback?.(rawResponse, flatResponse);
@@ -527,8 +524,8 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         ...args,
         options: {
           ...args.options,
-          onResponse: callback
-        }
+          onResponse: callback,
+        },
       };
       const flatResponse = await directSendOperation(updatedArgs, spec);
       return {
@@ -536,20 +533,20 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
         rawResponse: {
           statusCode: currentRawResponse!.status,
           body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
+          headers: currentRawResponse!.headers.toJSON(),
+        },
       };
     };
 
     const lro = createLroSpec({
       sendOperationFn,
       args: { resourceGroupName, profileName, originGroupName, options },
-      spec: deleteOperationSpec
+      spec: deleteOperationSpec,
     });
     const poller = await createHttpPoller<void, OperationState<void>>(lro, {
       restoreFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs,
-      resourceLocationConfig: "azure-async-operation"
+      resourceLocationConfig: "azure-async-operation",
     });
     await poller.poll();
     return poller;
@@ -567,13 +564,13 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     originGroupName: string,
-    options?: AfdOriginGroupsDeleteOptionalParams
+    options?: AfdOriginGroupsDeleteOptionalParams,
   ): Promise<void> {
     const poller = await this.beginDelete(
       resourceGroupName,
       profileName,
       originGroupName,
-      options
+      options,
     );
     return poller.pollUntilDone();
   }
@@ -590,11 +587,11 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     originGroupName: string,
-    options?: AfdOriginGroupsListResourceUsageOptionalParams
+    options?: AfdOriginGroupsListResourceUsageOptionalParams,
   ): Promise<AfdOriginGroupsListResourceUsageResponse> {
     return this.client.sendOperationRequest(
       { resourceGroupName, profileName, originGroupName, options },
-      listResourceUsageOperationSpec
+      listResourceUsageOperationSpec,
     );
   }
 
@@ -610,11 +607,11 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     resourceGroupName: string,
     profileName: string,
     nextLink: string,
-    options?: AfdOriginGroupsListByProfileNextOptionalParams
+    options?: AfdOriginGroupsListByProfileNextOptionalParams,
   ): Promise<AfdOriginGroupsListByProfileNextResponse> {
     return this.client.sendOperationRequest(
       { resourceGroupName, profileName, nextLink, options },
-      listByProfileNextOperationSpec
+      listByProfileNextOperationSpec,
     );
   }
 
@@ -632,11 +629,11 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
     profileName: string,
     originGroupName: string,
     nextLink: string,
-    options?: AfdOriginGroupsListResourceUsageNextOptionalParams
+    options?: AfdOriginGroupsListResourceUsageNextOptionalParams,
   ): Promise<AfdOriginGroupsListResourceUsageNextResponse> {
     return this.client.sendOperationRequest(
       { resourceGroupName, profileName, originGroupName, nextLink, options },
-      listResourceUsageNextOperationSpec
+      listResourceUsageNextOperationSpec,
     );
   }
 }
@@ -644,38 +641,15 @@ export class AfdOriginGroupsImpl implements AfdOriginGroups {
 const serializer = coreClient.createSerializer(Mappers, /* isXml */ false);
 
 const listByProfileOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups",
+  path: "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups",
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.AFDOriginGroupListResult
+      bodyMapper: Mappers.AFDOriginGroupListResult,
     },
     default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
-  },
-  queryParameters: [Parameters.apiVersion],
-  urlParameters: [
-    Parameters.$host,
-    Parameters.subscriptionId,
-    Parameters.resourceGroupName,
-    Parameters.profileName1
-  ],
-  headerParameters: [Parameters.accept],
-  serializer
-};
-const getOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
-  httpMethod: "GET",
-  responses: {
-    200: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AfdErrorResponse,
     },
-    default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
   },
   queryParameters: [Parameters.apiVersion],
   urlParameters: [
@@ -683,31 +657,51 @@ const getOperationSpec: coreClient.OperationSpec = {
     Parameters.subscriptionId,
     Parameters.resourceGroupName,
     Parameters.profileName1,
-    Parameters.originGroupName
   ],
   headerParameters: [Parameters.accept],
-  serializer
+  serializer,
+};
+const getOperationSpec: coreClient.OperationSpec = {
+  path: "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
+  httpMethod: "GET",
+  responses: {
+    200: {
+      bodyMapper: Mappers.AFDOriginGroup,
+    },
+    default: {
+      bodyMapper: Mappers.AfdErrorResponse,
+    },
+  },
+  queryParameters: [Parameters.apiVersion],
+  urlParameters: [
+    Parameters.$host,
+    Parameters.subscriptionId,
+    Parameters.resourceGroupName,
+    Parameters.profileName1,
+    Parameters.originGroupName,
+  ],
+  headerParameters: [Parameters.accept],
+  serializer,
 };
 const createOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
+  path: "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
   httpMethod: "PUT",
   responses: {
     200: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     201: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     202: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     204: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
+      bodyMapper: Mappers.AfdErrorResponse,
+    },
   },
   requestBody: Parameters.originGroup,
   queryParameters: [Parameters.apiVersion],
@@ -716,32 +710,31 @@ const createOperationSpec: coreClient.OperationSpec = {
     Parameters.subscriptionId,
     Parameters.resourceGroupName,
     Parameters.profileName1,
-    Parameters.originGroupName
+    Parameters.originGroupName,
   ],
   headerParameters: [Parameters.contentType, Parameters.accept],
   mediaType: "json",
-  serializer
+  serializer,
 };
 const updateOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
+  path: "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
   httpMethod: "PATCH",
   responses: {
     200: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     201: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     202: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     204: {
-      bodyMapper: Mappers.AFDOriginGroup
+      bodyMapper: Mappers.AFDOriginGroup,
     },
     default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
+      bodyMapper: Mappers.AfdErrorResponse,
+    },
   },
   requestBody: Parameters.originGroupUpdateProperties,
   queryParameters: [Parameters.apiVersion],
@@ -750,15 +743,14 @@ const updateOperationSpec: coreClient.OperationSpec = {
     Parameters.subscriptionId,
     Parameters.resourceGroupName,
     Parameters.profileName1,
-    Parameters.originGroupName
+    Parameters.originGroupName,
   ],
   headerParameters: [Parameters.contentType, Parameters.accept],
   mediaType: "json",
-  serializer
+  serializer,
 };
 const deleteOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
+  path: "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}",
   httpMethod: "DELETE",
   responses: {
     200: {},
@@ -766,8 +758,8 @@ const deleteOperationSpec: coreClient.OperationSpec = {
     202: {},
     204: {},
     default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
+      bodyMapper: Mappers.AfdErrorResponse,
+    },
   },
   queryParameters: [Parameters.apiVersion],
   urlParameters: [
@@ -775,22 +767,21 @@ const deleteOperationSpec: coreClient.OperationSpec = {
     Parameters.subscriptionId,
     Parameters.resourceGroupName,
     Parameters.profileName1,
-    Parameters.originGroupName
+    Parameters.originGroupName,
   ],
   headerParameters: [Parameters.accept],
-  serializer
+  serializer,
 };
 const listResourceUsageOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}/usages",
+  path: "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Cdn/profiles/{profileName}/originGroups/{originGroupName}/usages",
   httpMethod: "POST",
   responses: {
     200: {
-      bodyMapper: Mappers.UsagesListResult
+      bodyMapper: Mappers.UsagesListResult,
     },
     default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
+      bodyMapper: Mappers.AfdErrorResponse,
+    },
   },
   queryParameters: [Parameters.apiVersion],
   urlParameters: [
@@ -798,42 +789,21 @@ const listResourceUsageOperationSpec: coreClient.OperationSpec = {
     Parameters.subscriptionId,
     Parameters.resourceGroupName,
     Parameters.profileName1,
-    Parameters.originGroupName
+    Parameters.originGroupName,
   ],
   headerParameters: [Parameters.accept],
-  serializer
+  serializer,
 };
 const listByProfileNextOperationSpec: coreClient.OperationSpec = {
   path: "{nextLink}",
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.AFDOriginGroupListResult
+      bodyMapper: Mappers.AFDOriginGroupListResult,
     },
     default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
-  },
-  urlParameters: [
-    Parameters.$host,
-    Parameters.subscriptionId,
-    Parameters.resourceGroupName,
-    Parameters.profileName1,
-    Parameters.nextLink
-  ],
-  headerParameters: [Parameters.accept],
-  serializer
-};
-const listResourceUsageNextOperationSpec: coreClient.OperationSpec = {
-  path: "{nextLink}",
-  httpMethod: "GET",
-  responses: {
-    200: {
-      bodyMapper: Mappers.UsagesListResult
+      bodyMapper: Mappers.AfdErrorResponse,
     },
-    default: {
-      bodyMapper: Mappers.AfdErrorResponse
-    }
   },
   urlParameters: [
     Parameters.$host,
@@ -841,8 +811,29 @@ const listResourceUsageNextOperationSpec: coreClient.OperationSpec = {
     Parameters.resourceGroupName,
     Parameters.profileName1,
     Parameters.nextLink,
-    Parameters.originGroupName
   ],
   headerParameters: [Parameters.accept],
-  serializer
+  serializer,
+};
+const listResourceUsageNextOperationSpec: coreClient.OperationSpec = {
+  path: "{nextLink}",
+  httpMethod: "GET",
+  responses: {
+    200: {
+      bodyMapper: Mappers.UsagesListResult,
+    },
+    default: {
+      bodyMapper: Mappers.AfdErrorResponse,
+    },
+  },
+  urlParameters: [
+    Parameters.$host,
+    Parameters.subscriptionId,
+    Parameters.resourceGroupName,
+    Parameters.profileName1,
+    Parameters.nextLink,
+    Parameters.originGroupName,
+  ],
+  headerParameters: [Parameters.accept],
+  serializer,
 };