+++ sdk/cdn/arm-cdn/src/operationsInterfaces/routes.ts
@@ -18,7 +18,7 @@ import {
   RouteUpdateParameters,
   RoutesUpdateOptionalParams,
   RoutesUpdateResponse,
-  RoutesDeleteOptionalParams
+  RoutesDeleteOptionalParams,
 } from "../models";
 
 /// <reference lib="esnext.asynciterable" />
@@ -36,7 +36,7 @@ export interface Routes {
     resourceGroupName: string,
     profileName: string,
     endpointName: string,
-    options?: RoutesListByEndpointOptionalParams
+    options?: RoutesListByEndpointOptionalParams,
   ): PagedAsyncIterableIterator<Route>;
   /**
    * Gets an existing route with the specified route name under the specified subscription, resource
@@ -53,7 +53,7 @@ export interface Routes {
     profileName: string,
     endpointName: string,
     routeName: string,
-    options?: RoutesGetOptionalParams
+    options?: RoutesGetOptionalParams,
   ): Promise<RoutesGetResponse>;
   /**
    * Creates a new route with the specified route name under the specified subscription, resource group,
@@ -72,7 +72,7 @@ export interface Routes {
     endpointName: string,
     routeName: string,
     route: Route,
-    options?: RoutesCreateOptionalParams
+    options?: RoutesCreateOptionalParams,
   ): Promise<
     SimplePollerLike<OperationState<RoutesCreateResponse>, RoutesCreateResponse>
   >;
@@ -93,7 +93,7 @@ export interface Routes {
     endpointName: string,
     routeName: string,
     route: Route,
-    options?: RoutesCreateOptionalParams
+    options?: RoutesCreateOptionalParams,
   ): Promise<RoutesCreateResponse>;
   /**
    * Updates an existing route with the specified route name under the specified subscription, resource
@@ -112,7 +112,7 @@ export interface Routes {
     endpointName: string,
     routeName: string,
     routeUpdateProperties: RouteUpdateParameters,
-    options?: RoutesUpdateOptionalParams
+    options?: RoutesUpdateOptionalParams,
   ): Promise<
     SimplePollerLike<OperationState<RoutesUpdateResponse>, RoutesUpdateResponse>
   >;
@@ -133,7 +133,7 @@ export interface Routes {
     endpointName: string,
     routeName: string,
     routeUpdateProperties: RouteUpdateParameters,
-    options?: RoutesUpdateOptionalParams
+    options?: RoutesUpdateOptionalParams,
   ): Promise<RoutesUpdateResponse>;
   /**
    * Deletes an existing route with the specified route name under the specified subscription, resource
@@ -150,7 +150,7 @@ export interface Routes {
     profileName: string,
     endpointName: string,
     routeName: string,
-    options?: RoutesDeleteOptionalParams
+    options?: RoutesDeleteOptionalParams,
   ): Promise<SimplePollerLike<OperationState<void>, void>>;
   /**
    * Deletes an existing route with the specified route name under the specified subscription, resource
@@ -167,6 +167,6 @@ export interface Routes {
     profileName: string,
     endpointName: string,
     routeName: string,
-    options?: RoutesDeleteOptionalParams
+    options?: RoutesDeleteOptionalParams,
   ): Promise<void>;
 }