+++ sdk/core/core-client-rest/CHANGELOG.md
@@ -1,14 +1,20 @@
 # Release History
 
-## 1.3.2 (Unreleased)
+## 1.4.1 (Unreleased)
 
 ### Features Added
-
-- Support accept in headers.
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+### Other Changes
+
+## 1.4.0 (2024-04-09)
+
+### Features Added
+
+- Support accept in headers.
 
 ### Other Changes