+++ sdk/core/core-client/CHANGELOG.md
@@ -1,12 +1,20 @@
 # Release History
 
-## 1.9.2 (Unreleased)
+## 1.9.3 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+### Other Changes
+
+## 1.9.2 (2024-04-09)
+
+### Bugs Fixed
+
+- Address two issues related to Azure Disk Storage. [PR #29087](https://github.com/Azure/azure-sdk-for-js/pull/29087)
 
 ### Other Changes